Add missing support for paths and plugins
Previously, plugins (or rather, directories in `node_modules`), were
only checked globally and prefixed if the given value did not include
a path separator.  For example, when `foo/bar.js` was given, this
would not be looked up globally, or if a prefix was given (`baz`),
the prefixed version (`baz-foo/bar.js`) would not be checked either.